% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_submit.R
\name{buildExperimentEvent}
\alias{buildExperimentEvent}
\title{Event for when the experiment needs to be constructed}
\usage{
buildExperimentEvent(combined)
}
\description{
Event for when the experiment needs to be constructed
}
