% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_selection.R
\name{createAliquotSelectionTable}
\alias{createAliquotSelectionTable}
\title{Construct a data.frame for the aliquot selection table}
\usage{
createAliquotSelectionTable(exp, aliquotDf)
}
\description{
Construct a data.frame for the aliquot selection table
}
