% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_selection.R
\name{createCompoundSelectionTable}
\alias{createCompoundSelectionTable}
\title{Construct a data.frame for the compound selection table}
\usage{
createCompoundSelectionTable(exp)
}
\description{
Construct a data.frame for the compound selection table
}
