% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs_plots.R
\name{.outputConcentrationPlot}
\alias{.outputConcentrationPlot}
\title{Create a concentration plot in a Shiny application}
\usage{
.outputConcentrationPlot(input, output, exp)
}
\arguments{
\item{input}{The Shiny input object}

\item{output}{The Shiny output object}

\item{exp}{A reactive function that returns a SummarizedExperiment object
containing concentration data}
}
\description{
Sets up a plotly-based concentration plot in a Shiny application. This
function is designed to be called within a Shiny server function.
}
\details{
This function renders a concentration plot using `getConcentrationPlot`. It
requires that the provided SummarizedExperiment object has a
"concentration" assay. The function uses `shinyWidgets::execute_safely` to
handle potential errors gracefully.
}
