% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plots.R
\name{.toPlotly}
\alias{.toPlotly}
\title{Convert ggplot to plotly}
\usage{
.toPlotly(plot, dynamicTicks = TRUE)
}
\arguments{
\item{plot}{ggplot2 object}

\item{dynamicTicks}{Logical indicating whether to use dynamic ticks}
}
\value{
Plotly object of the given ggplot
}
\description{
This function can be used to convert any static plot
made with ggplot to an interactive plot with plotly. This is useful for
more complex plots, where panning and zooming may be needed.
}
\keyword{internal}
