% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plots.R
\name{getCompoundPlot}
\alias{getCompoundPlot}
\title{Create a compound plot}
\usage{
getCompoundPlot(
  exp,
  assay,
  batches,
  types,
  trendsTypes,
  columns = 2,
  logTransform = TRUE
)
}
\description{
Create a compound plot
}
