% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plots.R
\name{getConcentrationPlot}
\alias{getConcentrationPlot}
\title{Generate a concentration plot}
\usage{
getConcentrationPlot(exp, batches, types, plotOnCalibrationLine)
}
\arguments{
\item{exp}{A SummarizedExperiment object with concentration data}

\item{batches}{Character vector of batch names to include in the plot, or
"All" to include all batches}

\item{types}{Character vector specifying the types of concentration data to
plot}

\item{plotOnCalibrationLine}{Logical indicating whether to plot points on
the calibration line}
}
\value{
A plotly object representing the concentration plot
}
\description{
Creates a concentration plot based on a SummarizedExperiment object
containing concentration data.
}
\details{
This function creates a concentration plot using the `concentrationPlot`
function and converts it to a plotly object. If multiple batches are
specified, the plot is faceted by batch.
}
