% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_selection.R
\name{internalStandardPage}
\alias{internalStandardPage}
\title{Page the shows the internal standard selection tables}
\usage{
internalStandardPage()
}
\description{
Page the shows the internal standard selection tables
}
