% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observers_exports.R
\name{observeReportCreation}
\alias{observeReportCreation}
\title{Observers for creating reports and zip}
\usage{
observeReportCreation(input, exp)
}
\description{
Observers for creating reports and zip
}
