% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observers_submit.R
\name{observeSubmitEvent}
\alias{observeSubmitEvent}
\title{Oberserver for submitting data}
\usage{
observeSubmitEvent(session, input, aliquotDf, newExp)
}
\description{
Oberserver for submitting data
}
