% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observers_plots.R
\name{observeViolinPlotInputs}
\alias{observeViolinPlotInputs}
\title{Observe the inputs for the violin plot}
\usage{
observeViolinPlotInputs(input, exp)
}
\description{
Observe the inputs for the violin plot
}
