% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server.R
\name{openDashboard}
\alias{openDashboard}
\alias{shiny}
\title{Run mzQuality in the browser}
\usage{
openDashboard(
  browser = TRUE,
  host = "0.0.0.0",
  port = 3838,
  mount = NA,
  useLogin = FALSE,
  useDataLake = FALSE
)

shiny(
  browser = TRUE,
  host = "0.0.0.0",
  port = 3838,
  mount = NA,
  useLogin = FALSE,
  useDataLake = FALSE
)
}
\arguments{
\item{browser}{Should the browser be started?}

\item{host}{Which host should be used? Defaults to "127.0.0.0"}

\item{port}{which port number should be hosted}
}
\value{
Shiny application in a browser on port 3838
}
\description{
Run mzQuality in the browser
}
\examples{
# Don't run
if (FALSE) {
    openDashboard()
}
}
