% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs_plots.R
\name{outputBadQcPlot}
\alias{outputBadQcPlot}
\title{Create a QC violin plot in a Shiny application}
\usage{
outputBadQcPlot(input, output, exp)
}
\arguments{
\item{input}{The Shiny input object}

\item{output}{The Shiny output object}

\item{exp}{A reactive function that returns a SummarizedExperiment object}
}
\description{
Creates a QC violin plot in a Shiny application
}
\details{
This function renders a violin plot for QC samples
}
