% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs_standards.R
\name{outputCurrentStandardsTable}
\alias{outputCurrentStandardsTable}
\title{Create the output for the currently picked Internal Standards}
\usage{
outputCurrentStandardsTable(output, exp)
}
\description{
Create the output for the currently picked Internal Standards
}
