% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs_standards.R
\name{outputModifiedStandardsTable}
\alias{outputModifiedStandardsTable}
\title{Create the output for the optionally new Internal Standards}
\usage{
outputModifiedStandardsTable(input, output, exp)
}
\description{
Create the output for the optionally new Internal Standards
}
