% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs_plots.R
\name{outputPcaPlot}
\alias{outputPcaPlot}
\title{Create a PCA plot in a Shiny application}
\usage{
outputPcaPlot(input, output, exp)
}
\arguments{
\item{input}{The Shiny input object}

\item{output}{The Shiny output object}

\item{exp}{A reactive function that returns a SummarizedExperiment object}
}
\description{
Creates a principal component analysis plot
}
\details{
This function renders a PCA plot using renderPcaPlot
}
