% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs_plots.R
\name{outputRsdqcHeatmap}
\alias{outputRsdqcHeatmap}
\title{Create an RSDQC heatmap in a Shiny application}
\usage{
outputRsdqcHeatmap(input, output, exp)
}
\arguments{
\item{input}{The Shiny input object}

\item{output}{The Shiny output object}

\item{exp}{A reactive function that returns a SummarizedExperiment object}
}
\description{
Creates a heatmap of relative standard deviation for QC samples
}
\details{
This function renders an RSDQC heatmap plot
}
