% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs_plots.R
\name{outputSamplePlot}
\alias{outputSamplePlot}
\title{Create an aliquot plot in a Shiny application}
\usage{
outputSamplePlot(session, input, output, exp)
}
\arguments{
\item{input}{The Shiny input object}

\item{output}{The Shiny output object}

\item{exp}{A reactive function that returns a SummarizedExperiment object}
}
\description{
Creates an aliquot plot in a Shiny application
}
\details{
This function renders an aliquot plot using renderAliquotPlot
}
