% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilsServer.R
\name{toPlotly}
\alias{toPlotly}
\title{Convert ggplot to plotly}
\usage{
toPlotly(plot, dynamicTicks = TRUE)
}
\arguments{
\item{plot}{ggplot2 object}
}
\value{
Plotly object of the given ggplot
}
\description{
This function can be used to convert any static plot
made with ggplot to an interactive plot with plotly. This is useful for
more complex plots, where panning and zooming may be needed.
}
\examples{
# Read example dataset
exp <- readRDS(system.file(package = "mzQuality", "data.RDS"))

# Perform analysis
exp <- doAnalysis(exp)

# Create PCA plot
p <- pcaPlot(exp)

# Make the PCA interactive
toPlotly(p)
}
