% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_selection.R
\name{updateExperiment}
\alias{updateExperiment}
\title{Update the experiment when the qcType has changed}
\usage{
updateExperiment(input, exp, qcType)
}
\description{
Update the experiment when the qcType has changed
}
